split bucket sampler into own file, tests
